phase-v/services/audit-service/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.31.1
httpx==0.27.2
sqlmodel==0.0.22
psycopg[binary]==3.2.3
psycopg2-binary==2.9.10
pydantic==2.11.7
python-dotenv==1.0.1
